+++ trunk-recorder/plugin_manager/plugin_manager.cc
@@ -126,7 +126,7 @@ void plugman_poll_one() {
   }
 }
 
-void plugman_audio_callback(Recorder *recorder, float *samples, int sampleCount) {
+void plugman_audio_callback(Recorder *recorder, int16_t *samples, int sampleCount) {
   for (std::vector<Plugin *>::iterator it = plugins.begin(); it != plugins.end(); it++) {
     Plugin *plugin = *it;
     if (plugin->state == PLUGIN_RUNNING) {
